Add Glide for smooth and fast scroll.
add Glide to ListViewCostumeAdapter.java to load list of apps faster and smoother.
