+++ app/build.gradle
@@ -84,6 +84,10 @@ dependencies {
     implementation 'co.ronash.android:pushe-base:1.3.4'
     implementation 'com.google.android.gms:play-services-gcm:10.2.1'
 
+    //glide
+    compile "com.github.bumptech.glide:glide:3.7.0"
+
+
 
     testCompile 'junit:junit:4.12'
