+++ app/src/main/res/layout/detoxiom_widget_configure.xml
@@ -28,6 +28,7 @@
             android:id="@+id/list_view_conf_activity"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:scrollbars="vertical"
             android:visibility="gone">
 
         </android.support.v7.widget.RecyclerView>
