+++ app/src/main/res/layout/row_item.xml
@@ -5,13 +5,16 @@
     android:layout_gravity="center"
     android:gravity="center"
     android:orientation="vertical"
-    android:paddingBottom="5dp">
+    android:paddingBottom="8dp"
+    android:layout_margin="4dp"
+    android:paddingTop="8dp">
 
     <ImageView
         android:id="@+id/logoImage"
-        android:layout_width="75dp"
-        android:layout_height="75dp"
-        android:padding="20dp" />
+        android:layout_width="72dp"
+        android:layout_height="72dp"
+        android:padding="4dp"
+        android:scaleType="centerInside" />
 
     <TextView
         android:id="@+id/TitleText"
@@ -21,6 +24,7 @@
         android:ellipsize="end"
         android:gravity="center"
         android:maxLines="2"
+        android:padding="2dp"
         android:textAlignment="center"
         android:textSize="15sp"
         android:textStyle="bold" />
